+++ CHANGELOG.md
@@ -32,3 +32,7 @@
     - `LogManagerEngine` interface for logging systems communication layer
     - `LogManagementCoordinator` interface for log management changes coordinator
     - `LogManager` interface for log management operations
+- Built-in `LogManager` implementation
+    - `LogManagerImpl` default implementation
+- Built-in `LogManagementCoordinator` implementations
+    - `NoOpLogManagementCoordinator` local JVM only log management coordinator